Update index.js - Errors
Ook heb ik even in t document wat spaties vervangen met tabs

+++ index.js
@@ -6,9 +6,9 @@ const token = 'TOKEN';
 const PREFIX = ';';
 
 bot.on('ready', () =>{
-  console.log('This bot is online');
-  bot.user.setStatus('available')
-  bot.user.setActivity('Prefix ' + PREFIX)
+	console.log('This bot is online');
+	bot.user.setStatus('available')
+	bot.user.setActivity('Prefix ' + PREFIX)
 })
 
 // T is nu al een zooitje, maar ik maak hier n tijdelijke verzamelplek voor de embeds :)
@@ -17,7 +17,8 @@ const helpEmbed1 = new Discord.MessageEmbed()
 .setTitle('SprintJoekels Commands')
 .addFields(
  	{ name: `${PREFIX}help`, value: 'Displays this message, duh.' },
- 	{ name: `${PREFIX}ping`, value: 'Play a game of Ping-Pong' }
+ 	{ name: `${PREFIX}ping`, value: 'Play a game of Ping-Pong' },
+        { name: `${PREFIX}say`, value: 'Let the bot say something.'},
 )
 .setFooter('Showing page 1/1')
 
@@ -26,21 +27,25 @@ const helpEmbed1 = new Discord.MessageEmbed()
 bot.on('message', message => {
 	if (!message.content.startsWith(PREFIX) || message.author.bot) return;
 
- 	const args = message.content.slice(PREFIX.length).trim().split(/ +/);
- 	const command = args.shift().toLowerCase();
- 	const sayMessage = args.join(' ');
- 	const ping = Date.now() - message.createdTimestamp + ' ms';
+	const args = message.content.slice(PREFIX.length).trim().split(/ +/);
+	const command = args.shift().toLowerCase();
+	const sayMessage = args.join(' ');
+	const ping = Date.now() - message.createdTimestamp + ' ms';
 
- 	if (command === 'ping') {
-   	message.channel.send(`Pong! (${ping})`);
+	 if (command === 'ping') {
+   		message.channel.send(`Pong! (${ping})`);
  	}
- 	else if (command === 'help') {
-  	message.channel.send(helpEmbed1);
- 	}
-	else if (command === 'say') {
-   	message.channel.bulkDelete(1)
-  	message.channel.send(sayMessage)
-    }
+	else if (command === 'help') {
+ 		message.channel.send(helpEmbed1);
+   	else if (command === 'say') {
+		if (!args.length) {
+			 message.channel.send(`Error: The command ${PREFIX}say requires arguments. Example: ${PREFIX}say Test`);
+		}
+		else {
+			message.channel.bulkDelete(1)
+			message.channel.send(sayMessage)
+		}
+	}
 	// Gewoon hieronder wat andere commandjes toevoegen
 	else {
    	message.channel.send('Unknown command')
